Move activate_mutation and deactivate_mutation to Character (#35227)
