+++ src/npc.h
@@ -1028,6 +1028,11 @@ class npc : public player
         void execute_action( npc_action action ); // Performs action
         void process_turn() override;
 
+        using Character::invoke_item;
+        bool invoke_item( item *, const tripoint &pt ) override;
+        bool invoke_item( item *used, const std::string &method ) override;
+        bool invoke_item( item * ) override;
+
         /** rates how dangerous a target is from 0 (harmless) to 1 (max danger) */
         float evaluate_enemy( const Creature &target ) const;
