+++ src/player.h
@@ -897,25 +897,9 @@ class player : public Character
         void use( item_location loc );
         /** Uses the current wielded weapon */
         void use_wielded();
-        /**
-         * Asks how to use the item (if it has more than one use_method) and uses it.
-         * Returns true if it destroys the item. Consumes charges from the item.
-         * Multi-use items are ONLY supported when all use_methods are iuse_actor!
-         */
-        bool invoke_item( item *, const tripoint &pt );
-        /** As above, but with a pre-selected method. Debugmsg if this item doesn't have this method. */
-        bool invoke_item( item *, const std::string &, const tripoint &pt );
-        /** As above two, but with position equal to current position */
-        bool invoke_item( item * );
-        bool invoke_item( item *, const std::string & );
+
         /** Reassign letter. */
         void reassign_item( item &it, int invlet );
-
-        /** Consume charges of a tool or comestible item, potentially destroying it in the process
-         *  @param used item consuming the charges
-         *  @param qty number of charges to consume which must be non-zero
-         *  @return true if item was destroyed */
-        bool consume_charges( item &used, int qty );
 
         /** Removes gunmod after first unloading any contained ammo and returns true on success */
         bool gunmod_remove( item &gun, item &mod );
@@ -1363,17 +1347,8 @@ class player : public Character
         void print_encumbrance( const catacurses::window &win, int line = -1,
                                 const item *selected_clothing = nullptr ) const;
 
-        // Prints message(s) about current health
-        void print_health() const;
-
         using Character::query_yn;
         bool query_yn( const std::string &mes ) const override;
-
-        /**
-         * Has the item enough charges to invoke its use function?
-         * Also checks if UPS from this player is used instead of item charges.
-         */
-        bool has_enough_charges( const item &it, bool show_msg ) const;
 
         const pathfinding_settings &get_pathfinding_settings() const override;
         std::set<tripoint> get_path_avoid() const override;
@@ -1422,10 +1397,6 @@ class player : public Character
          * are included.
          */
         bool is_visible_in_range( const Creature &critter, int range ) const;
-
-        // Trigger and disable mutations that can be so toggled.
-        void activate_mutation( const trait_id &mutation );
-        void deactivate_mutation( const trait_id &mut );
 
         /** Determine player's capability of recharging their CBMs. */
         bool can_feed_reactor_with( const item &it ) const;
